+++ onmt/modules/init.lua
@@ -15,6 +15,8 @@ require('onmt.modules.LSTM')
 require('onmt.modules.MaskedSoftmax')
 require('onmt.modules.WordEmbedding')
 require('onmt.modules.FeaturesEmbedding')
+
+require('onmt.modules.NoAttention')
 require('onmt.modules.GlobalAttention')
 require('onmt.modules.GlobalAttentionCoverage')